+++ src/lib/pbkdf/pbkdf2/pbkdf2.cpp
@@ -178,7 +178,7 @@ PBKDF* PKCS5_PBKDF2::clone() const
 
 PBKDF2::PBKDF2(const MessageAuthenticationCode& prf, size_t olen, std::chrono::milliseconds msec) :
    m_prf(prf.clone()),
-   m_iterations(tune_pbkdf2(*m_prf, olen, msec.count()))
+   m_iterations(tune_pbkdf2(*m_prf, olen, static_cast<uint32_t>(msec.count())))
    {}
 
 std::string PBKDF2::to_string() const
